[Simplify Repo] 03 - Methods and Functions (#24)
* #23: create base js, test.js files
* Closes #23: extract problems into files; update descriptions, tests, command
