+++ lessons/03-Methods-and-Functions/01-helloWorld.js
@@ -1,0 +1,17 @@
+/**
+ * #1: helloWorld
+ *
+ * Define the function helloWorld. Make sure to export it for testing use.
+ * helloWorld takes no inputs and returns the phrase "Hello World!"
+ *
+ * @category 03 - Methods and Functions
+ * @function helloWorld
+ * @returns {String}
+ *
+ * @example
+ * helloWorld() => "Hello World!"
+ */
+
+export function helloWorld() {
+  return "Hello World!";
+}
