+++ lessons/03-Methods-and-Functions/02-helloWorldRedux.js
@@ -1,0 +1,30 @@
+/**
+ * #2: helloWorldRedux
+ *
+ * Define the function helloWorldRedux. Make sure to export it for testing use.
+ * helloWorldRedux takes in an optional input, name.
+ * When a name value is passed, return a personalized greeting using the name.
+ * Otherwise, return the phrase "Hello World!"
+ *
+ * CHALLENGE:
+ * Can you write this function as a one-liner? Remember to comment out the original function before testing your one-line version to avoid duplicate declaration errors!
+ *
+ * @category 03 - Methods and Functions
+ * @function helloWorldRedux
+ * @param {String?} name - optional name input
+ * @returns {String}
+ *
+ * @example
+ * helloWorldRedux() => "Hello World!"
+ * helloWorldRedux("Bob") => "Hello Bob!"
+ */
+
+export function helloWorldRedux(name) {
+  if (name) {
+    return `Hello ${name}!`;
+  }
+  return "Hello World!";
+}
+
+// One-line function
+// const helloWorldRedux = (name) => (name ? `Hello ${name}!` : `Hello World!`);
