+++ lessons/03-Methods-and-Functions/03-uppercaseThis.js
@@ -1,0 +1,30 @@
+/**
+ * #3: uppercaseThis
+ *
+ * Define the function uppercaseThis. Make sure to export it for testing use.
+ * uppercaseThis takes in one input, phrase, a String of any length.
+ * uppercaseThis returns phrase fully capitalized.
+ *
+ * CHALLENGE:
+ * Can you write this function as a one-liner? Remember to comment out the original function before testing your one-line version to avoid duplicate declaration errors!
+ *
+ * @category 03 - Methods and Functions
+ * @function uppercaseThis
+ * @param {String} phrase - string to be all-capped
+ * @returns {String} fully capitalized string
+ *
+ * @see {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/toUpperCase .toUpperCase()}
+ *
+ * @example
+ * uppercaseThis("aaaa") => "AAAA"
+ * uppercaseThis("BBBB") => "BBBB"
+ * uppercaseThis("cCcC") => "CCCC"
+ * uppercaseThis("d1D2d3") => "D1D2D3"
+ */
+
+export function uppercaseThis(phrase) {
+  return phrase.toUpperCase();
+}
+
+// One-line function
+// const uppercaseThis = phrase => phrase.toUpperCase()
